requirements.txt
tensorflow
opencv-python
Pillow
numpy
keras